hw_re/requirements.txt
re
requests
numpy
pandas
matplotlib.pyplot
